+++ lib/esbonio/esbonio/server/cli.py
@@ -70,6 +70,8 @@ def main(argv: Optional[Sequence[str]] = None):
         "esbonio.server.features.preview_manager",
         "esbonio.server.features.directives",
         "esbonio.server.features.roles",
+        "esbonio.server.features.rst.directives",
+        "esbonio.server.features.myst.directives",
         "esbonio.server.features.sphinx_support.diagnostics",
         "esbonio.server.features.sphinx_support.symbols",
         "esbonio.server.features.sphinx_support.directives",
